fix(profile): keep the running variance in double, not int64
Welford's update squares a deviation, and a zone's exclusive time is in
nanoseconds, so the product overflows a signed 64-bit integer as soon as the
spread reaches a few seconds: a sample 108 s from the mean squares to ~5.8e21
against a 9.2e18 ceiling. That is undefined behaviour, and the nightly
sanitizer leg reported it as exactly that - raised on the profiler's consumer
thread while draining at shutdown, which is what failed the CcsdTiled example
on that leg rather than anything in the example.

It only shows up under a sanitizer because that is where zones get slow enough,
but nothing about it is sanitizer-specific: the same multiply wraps in a
release build and reports a meaningless standard deviation.

The integer mean was wrong too, more quietly. Advancing it by `delta /
call_count` truncates every update, so it drifted low even when nothing
overflowed - for samples 10, 20, 30, 40 it reported 24 rather than 25.

Both accumulators are now double, which is the textbook formulation, and every
consumer already converted to double to use them.

The per-sample update moves to AggNode::record_exclusive so it can be exercised
directly: the durations that broke this are tens of seconds, and no test should
spend them.

+++ libs/Einsums/Profile/src/Profile.cpp
@@ -241,10 +241,8 @@ void Profiler::print_node_recursive(std::ostream &os, AggNode const *n, double /
     // frame size, so deep recursion exhausts stack even at modest tree
     // depth. Switching to an explicit work stack puts the per-node
     // bookkeeping on the heap and keeps a constant call-stack budget.
-    auto variance = [](uint64_t cnt, int64_t M2) -> double {
-        return (cnt > 1) ? static_cast<double>(M2) / static_cast<double>(cnt - 1) : 0.0;
-    };
-    auto stddev = [variance](uint64_t cnt, int64_t M2) -> double { return sqrt(variance(cnt, M2)); };
+    auto variance = [](uint64_t cnt, double M2) -> double { return (cnt > 1) ? M2 / static_cast<double>(cnt - 1) : 0.0; };
+    auto stddev   = [variance](uint64_t cnt, double M2) -> double { return sqrt(variance(cnt, M2)); };
 
     struct Frame {
         AggNode const *node;
@@ -265,7 +263,7 @@ void Profiler::print_node_recursive(std::ostream &os, AggNode const *n, double /
         if (name.size() > 60)
             name = name.substr(0, 57) + "...";
 
-        std::string const mean_str = fmt::format("{:7.3f}\u00B1{:3.3f}", static_cast<double>(node->total_exclusive_mean) / 1'000'000.0,
+        std::string const mean_str = fmt::format("{:7.3f}\u00B1{:3.3f}", node->total_exclusive_mean / 1'000'000.0,
                                                  stddev(node->call_count, node->total_exclusive_M2) / 1'000'000.0);
 
         // Build file:line field
